requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
telnyx==2.1.1
deepgram-sdk==3.2.7
groq==0.9.0
elevenlabs==1.2.2
google-api-python-client==2.131.0
google-auth==2.29.0
aiofiles==23.2.1
httpx==0.27.0
